+++ doc/version-history.rst
@@ -7,6 +7,8 @@ Version History
 v0.3.1
 ======
 * Add count check to trend_daily and hourly fields.
+* Attempt to pad data when count is mismatched.
+* Add simulation mode that has missing data.
 
 v0.3.0
 ======